runtime: move stdlib middleware to separate file
